+++ src/chapter_08_design_patterns/problem_071_observable_vector_container.cpp
@@ -16,7 +16,7 @@ void problem_71_main(std::ostream& os) {
     auto sp_ov_0{ std::make_shared<observable_vector<std::string>>() };
     auto sp_ov_1{ std::make_shared<observable_vector<float>>(5) };
     std::ranges::transform(*sp_ov_1, std::begin(*sp_ov_1),
-        [i = 0](auto&& f) mutable { return 3.14f * i++; }
+        [i = 0](auto&& f) mutable { return 3.14f * static_cast<float>(i++); }
     );
     auto sp_ov_2{ std::make_shared<observable_vector<double>>(3, 1.5) };
     const char* cstr{ "Hello, universe!" };
